requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
html2text==2024.2.26
langchain==0.3.14
langchain_community==0.3.14
langchain_core==0.3.29
langchain_experimental==0.3.4
langchain_openai==0.3.0
pydantic==2.10.5
pyfiglet==1.0.2
python-dotenv==1.0.1
rich==13.9.4
flask==3.0.0
flask-cors==4.0.0
# CTH dependencies
networkx==3.2.1
numpy>=1.24.0
scipy>=1.10.0